Addition of "main" file for Homestead Fl
I migrated all of the code from "BESSIE" and implemented it in a new main file which includes commented sections for variable assignment, computation, and graphing. It still generates the same graphs.
I also added a plot for the overloads if we upgraded the substation by 10MW.
A few variables had name changes which I felt made more sense.
Eg. netLoad -> netLoadSolar