RealBESStFunc WORKS!
ITS ALLIIIVVEEE!  The time-based battery charge and discharge function is working properly, however there are still some things that could use improving such as time for charging via solar and for when overloads occur during designated discharging times.  There are also two BESSt functions that are not used in the main file (BESStFunc and BESSt2Func) while the RealBESStFunc is the actual code.  Both of the others are notes and other ideas, so I didn't want to delete them.